# Improve compilation performance