--- a/src/keywords/any_of.rs
+++ b/src/keywords/any_of.rs
@@ -12,17 +12,15 @@ pub struct AnyOfValidator {
 impl AnyOfValidator {
     #[inline]
     pub(crate) fn compile(schema: &Value, context: &CompilationContext) -> CompilationResult {
-        match schema.as_array() {
-            Some(items) => {
-                let mut schemas = Vec::with_capacity(items.len());
-                for item in items {
-                    let validators = compile_validators(item, context)?;
-                    schemas.push(validators)
-                }
-                Ok(Box::new(AnyOfValidator { schemas }))
+        if let Value::Array(items) = schema {
+            let mut schemas = Vec::with_capacity(items.len());
+            for item in items {
+                let validators = compile_validators(item, context)?;
+                schemas.push(validators)
             }
-            None => Err(CompilationError::SchemaError),
+            return Ok(Box::new(AnyOfValidator { schemas }));
         }
+        Err(CompilationError::SchemaError)
     }
 }