--- a/src/keywords/legacy/maximum_draft_4.rs
+++ b/src/keywords/legacy/maximum_draft_4.rs
@@ -2,6 +2,7 @@ use super::super::{exclusive_maximum, maximum, CompilationResult};
 use crate::compilation::CompilationContext;
 use serde_json::{Map, Value};
 
+#[inline]
 pub(crate) fn compile(
     parent: &Map<String, Value>,
     schema: &Value,