--- a/src/keywords/max_properties.rs
+++ b/src/keywords/max_properties.rs
@@ -10,6 +10,7 @@ pub struct MaxPropertiesValidator {
 }
 
 impl<'a> MaxPropertiesValidator {
+    #[inline]
     pub(crate) fn compile(schema: &Value) -> CompilationResult {
         if let Value::Number(limit) = schema {
             let limit = limit.as_u64().unwrap() as usize;
@@ -42,6 +43,8 @@ impl Validate for MaxPropertiesValidator {
         format!("<max properties: {}>", self.limit)
     }
 }
+
+#[inline]
 pub(crate) fn compile(
     _: &Map<String, Value>,
     schema: &Value,