@Stranger6667 2020-05-20
So, for some more or less realistic schemas, this PR gives 8-12% improvement in the compilation time and from 40% to 80% improvement for small single-keyword schemas